backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-jose[cryptography]
passlib[bcrypt]